+++ .gitignore
@@ -23,4 +23,10 @@ target/
 # Python / maturin files
 #  These are generated by maturin and contain build artifacts
 __pycache__/
+*.py[cod]
+*$py.class
+build/
+dist/
+*.egg-info/
+*.so
