+++ crates/serpen/src/lib.rs
@@ -4,6 +4,7 @@ pub mod dependency_graph;
 pub mod emit;
 pub mod python_stdlib;
 pub mod resolver;
+pub mod util;
 
 pub use bundler::Bundler;
 pub use config::Config;
